# Improve some thermo calculation bottlenecks by refactoring how units are handled