--- a/setup.cfg
+++ b/setup.cfg
@@ -88,7 +88,7 @@ per-file-ignores = examples/*.py: D MPY001 T003 T001
                    src/metpy/interpolate/*.py: RST306
                    src/metpy/io/*.py: RST306
                    src/metpy/future.py: RST307
-                   src/metpy/constants.py: RST306
+                   src/metpy/constants/*.py: RST306
                    docs/doc-server.py: T001
                    tests/*.py: MPY001
                    ci/filter_links.py: E731 T001