#### Description Of Changes

I noticed that there have been two major performance bottlenecks caused by the existing approaches to unit handling: iteration in `moist_lapse`/`lcl` (#1169) and looping of `thickness_hydrostatic` in `io.gempak`'s `_interp_*` functions (#2062). In an attempt to improve this situation, this PR takes the approach of refactoring the relevant thermo calculations to have private versions that only use base units (no `pint.Quantity` handling), and wrapping those while handling units. No robust benchmarks were evaluated, but this did speed up `pytest tests/calc/test_thermo.py` on my workstation from about 3.5 seconds to 0.69 seconds (similar to #1980), and @akrherz's test file in https://github.com/Unidata/MetPy/issues/2041#issuecomment-904920081 from 35.88 seconds to 1.78 seconds.

I tried to be careful to not touch any tests in this refactor (other than the fix to the flake8 checker), but do let me know if any tests should be added.

Future work building on this PR could examine if all/most of the calculations should be refactored in this way (which may also reveal some cleaner implementation approaches) and if numba `jit`/`guvectorize` could meaningfully accelerate any of these private routines.

@sgdecker, could you evaluate how much performance improvement this gives you relative to your earlier tests in #2062?

@nawendt, these changes caused some failing tests in the GEMPAK reader due to what looks to be marginally different output values. Would you be willing to take a look and see if these differences are significant or not?

#### Checklist

- [x] Closes #1169, closes #2062
- [ ] Tests added
- [ ] Fully documented