@nawendt 2021-08-29
I added a comment to `convert_degC_to_K`. This function is the reason the tests started failing. It did not account for missing values which were most likely -9999. The reference data has a `nan` value whereas the decoded soundings ended up having some nonsense values in the same array position. Glad it was an easy fix.