@sgdecker 2021-08-30
On an old laptop, I am going from 3:00 to 0:13 with this branch and my test file. Very nice!