@jthielen 2021-08-31
I wasn't all that happy with the private routine + public wrapper approach...it would get really messy if applied to the whole calculation suite. So, I went back and came up with a decorator-based alternative: https://gist.github.com/jthielen/93ec9fcd8f3892a83215309fe5cda2c6. Doing it this way moves all unit specification into the decorator (which does carry with it the slight issue of default arguments in the signature needing to be in base units and not Quanitities), and leaves the unitless version attached to the parent function on the `._nounit` attribute (for use internally).

The existing approach is passing all checks, but let me know if it'd be useful and go back and redo it in the decorator-based way. At the very least, I'd want to clean it up to allow the existing `check_units` and this new `process_units` to better share code, and to think through if I'm missing any edge cases in the design. Otherwise, if we want to get these performance improvements in as soon as possible, the decorator alternative can be left to a future refactor when we need to optimize more routines.