@jthielen 2021-09-17
This has now all been rewritten to use the decorator-based approach! Performance improvements from my tests are about the same as before.