@jthielen 2022-01-14
This should be ready for final review once it is adapted after the merge of https://github.com/Unidata/MetPy/pull/2263