@jthielen 2022-01-18
@dcamron I've rebased this on main now that https://github.com/Unidata/MetPy/pull/2263 is in. I'm not on my primary workstation at the moment, so wasn't able to check full lint, test suite, and performance, but at least we can see what CI thinks. Also, should be good enough for a review.