@dopplershift 2022-01-18
We can ignore Code Climate.